Fixing warnings

+++ Source/DREAM3DLib/GenericFilters/FindNeighbors.cpp
@@ -158,7 +158,7 @@ void FindNeighbors::execute()
 
 
   int64_t totalPoints = m->getTotalPoints();
-  int totalFields = m->getNumFieldTuples();
+  int totalFields = int(m->getNumFieldTuples());
   dataCheck(false, totalPoints, totalFields, m->getNumEnsembleTuples());
   if (getErrorCondition() < 0)
   {
@@ -209,7 +209,7 @@ void FindNeighbors::execute()
   for (int i = 1; i < totalFields; i++)
   {
     std::stringstream ss;
-    ss << "Finding Neighbors - Initializing Neighbor Lists - " << ((float)i/totalFields)*100 << " Percent Complete";
+    ss << "Finding Neighbors - Initializing Neighbor Lists - " << (static_cast<float>(i)/totalFields)*100 << " Percent Complete";
  //   notifyProgress(ss.str(), 0, Observable::UpdateProgressMessage);
     m_NumNeighbors[i] = 0;
     neighborlist[i].resize(nListSize);
@@ -222,15 +222,15 @@ void FindNeighbors::execute()
   for (int64_t j = 0; j < totalPoints; j++)
   {
     std::stringstream ss;
-    ss << "Finding Neighbors - Determining Neighbor Lists - " << ((float)j/totalPoints)*100 << " Percent Complete";
+    ss << "Finding Neighbors - Determining Neighbor Lists - " << (static_cast<float>(j)/totalPoints)*100 << " Percent Complete";
  //   notifyProgress(ss.str(), 0, Observable::UpdateProgressMessage);
     onsurf = 0;
     grain = m_GrainIds[j];
     if(grain > 0)
     {
-      column = j % m->getXPoints();
-      row = (j / m->getXPoints()) % m->getYPoints();
-      plane = j / (m->getXPoints() * m->getYPoints());
+      column = static_cast<float>( j % m->getXPoints() );
+      row = static_cast<float>( (j / m->getXPoints()) % m->getYPoints() );
+      plane = static_cast<float>( j / (m->getXPoints() * m->getYPoints()) );
       if((column == 0 || column == (m->getXPoints() - 1) || row == 0 || row == (m->getYPoints() - 1) || plane == 0 || plane == (m->getZPoints() - 1)) && m->getZPoints() != 1)
       {
         m_SurfaceFields[grain] = true;
@@ -242,7 +242,7 @@ void FindNeighbors::execute()
       for (int k = 0; k < 6; k++)
       {
         good = 1;
-        neighbor = j + neighpoints[k];
+        neighbor = static_cast<int>( j + neighpoints[k] );
         if(k == 0 && plane == 0) good = 0;
         if(k == 5 && plane == (m->getZPoints() - 1)) good = 0;
         if(k == 1 && row == 0) good = 0;
@@ -255,7 +255,7 @@ void FindNeighbors::execute()
           nnum = m_NumNeighbors[grain];
           neighborlist[grain].push_back(m_GrainIds[neighbor]);
           nnum++;
-          m_NumNeighbors[grain] = nnum;
+          m_NumNeighbors[grain] = static_cast<int32_t>(nnum);
         }
       }
     }
@@ -273,7 +273,7 @@ void FindNeighbors::execute()
     int phase = m_FieldPhases[i];
 
     std::map<int, int> neighToCount;
-    int numneighs = int(neighborlist[i].size());
+    int numneighs = static_cast<int>( neighborlist[i].size() );
 
     // this increments the voxel counts for each grain
     for (int j = 0; j < numneighs; j++)
@@ -301,18 +301,17 @@ void FindNeighbors::execute()
       neighborlist[i].push_back(neigh);
       neighborsurfacearealist[i].push_back(area);
     }
-    m_NumNeighbors[i] = neighborlist[i].size();
+    m_NumNeighbors[i] = int32_t( neighborlist[i].size() );
 
     // Set the vector for each list into the NeighborList Object
     NeighborList<int>::SharedVectorType sharedNeiLst(new std::vector<int>);
     sharedNeiLst->assign(neighborlist[i].begin(), neighborlist[i].end());
-    m_NeighborList->setList(i, sharedNeiLst);
+    m_NeighborList->setList(static_cast<int>(i), sharedNeiLst);
 
     NeighborList<float>::SharedVectorType sharedSAL(new std::vector<float>);
     sharedSAL->assign(neighborsurfacearealist[i].begin(), neighborsurfacearealist[i].end());
-    m_SharedSurfaceAreaList->setList(i, sharedSAL);
+    m_SharedSurfaceAreaList->setList(static_cast<int>(i), sharedSAL);
   }
 
  notifyProgress("Finding Neighbors Complete", 0, Observable::UpdateProgressMessage);
 }
-
